+++ lib/screens/connect/connect_screen.dart
@@ -1,9 +1,10 @@
 import 'package:clipboard/clipboard.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-import 'package:sentry_flutter/sentry_flutter.dart';
+
 import 'package:url_launcher/url_launcher.dart';
 
+import '../../mixin/sentry_state_transaction_mixin.dart';
 import '../../redux/state/app_state.dart';
 import '../../screens/scanner/scanner_screen.dart';
 import '../../utils/sentry_colors.dart';
@@ -14,23 +15,8 @@ class ConnectScreen extends StatefulWidget {
   _ConnectScreenState createState() => _ConnectScreenState();
 }
 
-class _ConnectScreenState extends State<ConnectScreen> {
+class _ConnectScreenState extends State<ConnectScreen> with SentryStateTransactionMixin<ConnectScreen> {
   var _loading = false;
-
-  ISentrySpan? transaction;
-
-  @override
-  void initState() {
-    super.initState();
-    transaction = Sentry.startTransaction(
-      '_ConnectScreenState',
-      'ui.load',
-      bindToScope: true,
-    );
-    WidgetsBinding.instance?.addPostFrameCallback((_){
-      transaction?.finish();
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
